+++ access-token-management/src/AccessTokenManagement.OpenIdConnect/AuthenticateResultCache.cs
@@ -2,10 +2,12 @@
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
 using Microsoft.AspNetCore.Authentication;
+
+namespace Duende.AccessTokenManagement.OpenIdConnect;
+
 /// <summary>
 /// Per-request cache so that if SignInAsync is used, we won't re-read the old/cached AuthenticateResult from the handler.
 /// This requires this service to be added as scoped to the DI system.
 /// Be VERY CAREFUL to not accidentally capture this service for longer than the appropriate DI scope - e.g., in an HttpClient.
 /// </summary>
-// ReSharper disable once CheckNamespace
 internal class AuthenticateResultCache : Dictionary<string, AuthenticateResult>;
