+++ access-token-management/test/AccessTokenManagement.Tests/PublicApiVerificationTests.GetAllPublicTypes.verified.txt
@@ -1,6 +1,4 @@
-﻿Microsoft.Extensions.DependencyInjection.ClientCredentialsTokenManagementBuilder
-Microsoft.Extensions.DependencyInjection.ClientCredentialsTokenManagementServiceCollectionExtensions
-Duende.AccessTokenManagement.AccessTokenHandler
+﻿Duende.AccessTokenManagement.AccessTokenHandler (obsolete)
 Duende.AccessTokenManagement.ActivitySources
 Duende.AccessTokenManagement.ActivitySourceNames
 Duende.AccessTokenManagement.ActivityNames
@@ -8,16 +6,17 @@ Duende.AccessTokenManagement.ClientCredentialsClient
 Duende.AccessTokenManagement.ClientCredentialsToken
 Duende.AccessTokenManagement.ClientCredentialsTokenEndpointService (obsolete)
 Duende.AccessTokenManagement.ClientCredentialsTokenHandler (obsolete)
+Duende.AccessTokenManagement.ClientCredentialsTokenManagementBuilder
 Duende.AccessTokenManagement.ClientCredentialsTokenManagementDefaults
 Duende.AccessTokenManagement.ClientCredentialsTokenManagementOptions
 Duende.AccessTokenManagement.ClientCredentialsTokenManagementService (obsolete)
+Duende.AccessTokenManagement.ClientCredentialsTokenManagementServiceCollectionExtensions
 Duende.AccessTokenManagement.DefaultClientAssertionService (obsolete)
 Duende.AccessTokenManagement.DefaultDPoPKeyStore (obsolete)
 Duende.AccessTokenManagement.DefaultDPoPProofService (obsolete)
 Duende.AccessTokenManagement.DistributedClientCredentialsTokenCache (obsolete)
 Duende.AccessTokenManagement.DistributedDPoPNonceStore (obsolete)
 Duende.AccessTokenManagement.DPoPExtensions
-Duende.AccessTokenManagement.DuendeAccessTokenSerializationContext
 Duende.AccessTokenManagement.IClientAssertionService
 Duende.AccessTokenManagement.IClientCredentialsCacheKeyGenerator
 Duende.AccessTokenManagement.IClientCredentialsTokenCache
@@ -34,4 +33,4 @@ Duende.AccessTokenManagement.DPoPProof
 Duende.AccessTokenManagement.ITokenRequestSynchronization
 Duende.AccessTokenManagement.ServiceProviderKeys
 Duende.AccessTokenManagement.TokenRequestParameters
-Duende.AccessTokenManagement.OTel.AccessTokenManagementMetrics
+Duende.AccessTokenManagement.OTel.AccessTokenManagementMetrics (obsolete)
