+++ access-token-management/test/AccessTokenManagement.Tests/PublicApiVerificationTests.GetAllPublicTypes_OpenIdConnect.verified.txt
@@ -1,6 +1,4 @@
-﻿Microsoft.AspNetCore.Authentication.TokenManagementHttpContextExtensions
-Microsoft.Extensions.DependencyInjection.OpenIdConnectTokenManagementServiceCollectionExtensions
-Duende.AccessTokenManagement.OpenIdConnect.AuthenticationSessionUserAccessTokenStore (obsolete)
+﻿Duende.AccessTokenManagement.OpenIdConnect.AuthenticationSessionUserAccessTokenStore (obsolete)
 Duende.AccessTokenManagement.OpenIdConnect.BlazorServerUserAccessor (obsolete)
 Duende.AccessTokenManagement.OpenIdConnect.CircuitServicesAccessor (obsolete)
 Duende.AccessTokenManagement.OpenIdConnect.ConfigureOpenIdConnectClientCredentialsOptions
@@ -17,8 +15,10 @@ Duende.AccessTokenManagement.OpenIdConnect.OpenIdConnectClientAccessTokenHandler
 Duende.AccessTokenManagement.OpenIdConnect.OpenIdConnectClientConfiguration
 Duende.AccessTokenManagement.OpenIdConnect.OpenIdConnectConfigurationService (obsolete)
 Duende.AccessTokenManagement.OpenIdConnect.OpenIdConnectTokenManagementDefaults
+Duende.AccessTokenManagement.OpenIdConnect.OpenIdConnectTokenManagementServiceCollectionExtensions
 Duende.AccessTokenManagement.OpenIdConnect.OpenIdConnectUserAccessTokenHandler (obsolete)
 Duende.AccessTokenManagement.OpenIdConnect.StoreTokensInAuthenticationProperties (obsolete)
+Duende.AccessTokenManagement.OpenIdConnect.TokenManagementHttpContextExtensions
 Duende.AccessTokenManagement.OpenIdConnect.UserAccessAccessTokenManagementService (obsolete)
 Duende.AccessTokenManagement.OpenIdConnect.UserToken
 Duende.AccessTokenManagement.OpenIdConnect.UserTokenEndpointService (obsolete)
